eliminate grpc paths
