+++ packages/insomnia/src/ui/components/dropdowns/grpc-method-dropdown/grpc-method-dropdown.tsx
@@ -1,13 +1,7 @@
 import React, { Fragment, FunctionComponent } from 'react';
 import styled from 'styled-components';
 
-import {
-  getGrpcPathSegments,
-  getShortGrpcPath,
-  groupGrpcMethodsByPackage,
-  GrpcMethodInfo,
-  NO_PACKAGE_KEY,
-} from '../../../../common/grpc-paths';
+import type { GrpcMethodInfo } from '../../../../main/ipc/grpc';
 import { Dropdown } from '../../base/dropdown/dropdown';
 import { DropdownButton } from '../../base/dropdown/dropdown-button';
 import { DropdownDivider } from '../../base/dropdown/dropdown-divider';
@@ -41,7 +35,36 @@ interface Props {
   handleChange: (arg0: string) => void;
   handleChangeProtoFile: () => void;
 }
+const PROTO_PATH_REGEX = /^\/(?:(?<package>[\w.]+)\.)?(?<service>\w+)\/(?<method>\w+)$/;
 
+export const NO_PACKAGE_KEY = 'no-package';
+
+function groupBy(list: {}[], keyGetter: (item: any) => string):Record<string, any[]> {
+  const map = new Map();
+  list.forEach(item => {
+    const key = keyGetter(item);
+    const collection = map.get(key);
+    if (!collection) {
+      map.set(key, [item]);
+    } else {
+      collection.push(item);
+    }
+  });
+  return Object.fromEntries(map);
+}
+
+export const groupGrpcMethodsByPackage = (methodInfoList: GrpcMethodInfo[]): Record<string, GrpcMethodInfo[]> => {
+  return groupBy(methodInfoList, ({ segments }) => segments.packageName || NO_PACKAGE_KEY);
+};
+
+// If all segments are found, return a shorter path, otherwise the original path
+export const getShortGrpcPath = (fullPath: string): string => {
+  const result = PROTO_PATH_REGEX.exec(fullPath);
+  const packageName = result?.groups?.package;
+  const serviceName = result?.groups?.service;
+  const methodName = result?.groups?.method;
+  return packageName && serviceName && methodName ? `/${serviceName}/${methodName}` : fullPath;
+};
 const NormalCase = styled.span`
   text-transform: initial;
 `;
@@ -64,7 +87,7 @@ export const GrpcMethodDropdown: FunctionComponent<Props> = ({
         buttonClass={DropdownMethodButton}
       >
         <Tooltip message={selectedPath || 'Select Method'} position="bottom" delay={500}>
-          {!selectedPath ? 'Select Method' : getShortGrpcPath(getGrpcPathSegments(selectedPath), selectedPath)}
+          {!selectedPath ? 'Select Method' : getShortGrpcPath(selectedPath)}
           <i className="fa fa-caret-down pad-left-sm" />
         </Tooltip>
       </DropdownButton>
@@ -82,7 +105,7 @@ export const GrpcMethodDropdown: FunctionComponent<Props> = ({
           <DropdownDivider>
             {name !== NO_PACKAGE_KEY && <NormalCase>pkg: {name}</NormalCase>}
           </DropdownDivider>
-          {pkg.map(({ segments, type, fullPath }) => (
+          {pkg.map(({ type, fullPath }) => (
             <DropdownItem
               key={fullPath}
               onClick={() => handleChange(fullPath)}
@@ -90,7 +113,7 @@ export const GrpcMethodDropdown: FunctionComponent<Props> = ({
               selected={fullPath === selectedPath}
             >
               <Tooltip message={fullPath} position="right" delay={500}>
-                <DropdownMethodButtonLabel><GrpcMethodTag methodType={type} /> {getShortGrpcPath(segments, fullPath)}</DropdownMethodButtonLabel>
+                <DropdownMethodButtonLabel><GrpcMethodTag methodType={type} /> {getShortGrpcPath(fullPath)}</DropdownMethodButtonLabel>
               </Tooltip>
             </DropdownItem>
           ))}
